stored procedures
there is list of changes from arturo in checkout page as coments, feel free to do any of them

+++ Models/ObjectOfIds.cs
@@ -8,7 +8,7 @@ namespace database_final_project.Models
     public class ObjectOfFieldsForDatabase
     {
         public int UserId { get; set; }
-        public int CreditCardId { get; set; }
+        public Int64 CreditCardId { get; set; }
         public string Products { get; set; }
         public decimal Total { get; set; }
         public decimal Tax { get; set; }
